requirements.txt
Django==2.2.17
dnspython